refactor : 코드 리뷰 반영(mixpanel 오타 수정)
