--- src/main/kotlin/com/deepromeet/atcha/miaxpanel/event/MixpanelEventProperty.kt
@@ -1,3 +1,0 @@
-package com.deepromeet.atcha.miaxpanel.event
-
-interface MixpanelEventProperty
