+++ src/main/kotlin/com/deepromeet/atcha/mixpanel/event/BusApiCallCountPerRequestProperty.kt
@@ -1,4 +1,4 @@
-package com.deepromeet.atcha.miaxpanel.event
+package com.deepromeet.atcha.mixpanel.event
 
 class BusApiCallCountPerRequestProperty(
     var oDsayCallCount: Int = 0,
