+++ src/main/kotlin/com/deepromeet/atcha/mixpanel/event/MixpanelEventProperty.kt
@@ -1,0 +1,3 @@
+package com.deepromeet.atcha.mixpanel.event
+
+interface MixpanelEventProperty
