+++ src/main/kotlin/com/deepromeet/atcha/transit/application/bus/BusScheduleProvider.kt
@@ -1,6 +1,6 @@
 package com.deepromeet.atcha.transit.application.bus
 
-import com.deepromeet.atcha.miaxpanel.event.BusApiCallCountPerRequestProperty
+import com.deepromeet.atcha.mixpanel.event.BusApiCallCountPerRequestProperty
 import com.deepromeet.atcha.transit.domain.bus.BusRouteInfo
 import com.deepromeet.atcha.transit.domain.bus.BusSchedule
